fix: не работает метод поиска с фильтрацией событий

+++ main-server/src/main/java/ru/practicum/app/controllers/PublicController.java
@@ -84,7 +84,7 @@ public class PublicController {
     @GetMapping(value = "/events")
     public Collection<EventShortDto> getFilteredEvents(
             @RequestParam(required = false) String text,
-            @RequestParam(required = false) List<Long> categories,
+            @RequestParam(required = false) List<Integer> categories,
             @RequestParam(required = false) Boolean paid,
             @RequestParam(required = false) String rangeStart,
             @RequestParam(required = false) String rangeEnd,
